uncomment baseUrl line  in tsconfig and fix related errors
